행사 일정 (시작일 종료일) 삽입
제목같은건 tour_content에서 끌어오면될듯

+++ DUGAZA/src/main/java/kr/spring/api/client/TourApiClient.java
@@ -21,7 +21,7 @@ public class TourApiClient {
     private static final DateTimeFormatter DATE_TIME_FORMATTER = DateTimeFormatter.ofPattern("yyyyMMddHHmmss");
 
     @LogExecutionTime(category = "TourData")
-    public List<TourApiDto> getTouristData(int contentTypeId) {
+    public List<TourApiDto> getTouristData(Long contentTypeId) {
         URI uri = baseApiClient.makeTourUri("/areaBasedList2","contentTypeId", String.valueOf(contentTypeId));
         return baseApiClient.callApiManyTimes(uri, this::createTourApiDto);
     }
